fix(pricing): the chain's last two integer revision columns become bigint — Z6-7
A plan revision is a `u64` wherever it is a value, and every column in the chain
that carries one is `bigint` — except `pricing_migration.source_revision` and
`pricing_snapshot_provenance.source_revision`, which were `integer`. Guarded
rather than broken: both writers narrowed with `i32::try_from` and answered
`CorruptRow`, so the consequence was a fail-closed refusal of a revision the
plan's own column can hold. Invisible from the entity too — it said `i32` and a
reader had to reach the DDL.

**A rebuild had already passed over one of them.** `m20260802_000065` rebuilt
`pricing_migration` on SQLite to widen its key by tenant and restated
`source_revision integer NOT NULL` verbatim. That is the shape worth naming: the
cheapest possible moment to correct a column is a migration that re-types the
whole table, and a verbatim restatement carries the outlier forward invisibly.

Postgres does the work — two `ALTER COLUMN … TYPE bigint`, no rebuild, no rows
touched, no object restated; the two CHECKs hold over the wider type unchanged.
**The SQLite arm is empty on purpose and the migration argues it rather than
skipping it**: on that engine `integer` and `bigint` denote one type (INTEGER
affinity, up to 8 bytes), so there was never a narrowing there to widen and the
only thing an arm could change is a word in `sqlite_master` — bought with a full
table rebuild of three indexes and five triggers. The `down` narrows and can fail
on a row written while the column was wide; that is stated as correct rather than
worked around, because `USING …::integer` would silently truncate a schedule's
source pointer.

Both `i32::try_from`s become `i64::try_from` — the column's own range rather than
a third one — and both entity fields become `i64` / `Option<i64>`.

Three probes, RED first against **all four sites reverted at once** rather than
one at a time:

- `sqlite_migration_repo::a_revision_beyond_the_old_columns_range_round_trips`
- `sqlite_snapshot_provenance_store::a_revision_beyond_the_old_columns_range_round_trips`,
  the one case in a deliberately raw-SQL suite that goes through the repository,
  and the suite doc says why: no raw statement can distinguish the column before
  from the column after on that engine, so a raw probe would have been green
  against the defect. What was narrow was the Rust.
  Both use `i32::MAX as u64 + 1` — the old boundary exactly, because one less
  passed before the widening too — and both read back rather than only inserting.
- `postgres_migrations::every_revision_column_is_bigint` states the property over
  the **schema**, not over these two columns: every `…_revision` column in `bss`,
  off `information_schema`, so the next table that types a revision `integer`
  reddens here. A spot check on the known outliers would be green against the
  third. RED with the migration's Postgres statements emptied.

Postgres tier EXECUTED: `postgres_migrations` 10 -> 11 green under `-- --ignored`,
and `postgres_schema_migration` 11 / `postgres_schema_snapshot_provenance` 18 /
`postgres_audit_chain` 11 re-run green against the wider columns. Fast tier:
`sqlite_migration_repo` 18 -> 19, `sqlite_snapshot_provenance_store` 9 -> 10.

+++ gears/bss/pricing/pricing/src/infra/storage/entity/migration.rs
@@ -43,7 +43,12 @@ pub struct Model {
     /// The retiring side.
     pub source_plan_id: Uuid,
     /// The source plan's revision at schedule time.
-    pub source_revision: i32,
+    ///
+    /// `i64` since `m20260802_000075` (Z6-7). It was `i32` against an `integer`
+    /// column, which made this one of two revision columns in the chain narrower
+    /// than the `u64` a revision is — visible from the DDL only, which is why the
+    /// review found it and no reader had.
+    pub source_revision: i64,
     /// MUST be published when the schedule is created, and re-checked nowhere:
     /// a target that stops being published afterwards surfaces as a D-36
     /// execution-time exclusion rather than as a retroactive refusal.
